Added support for progress forum links.
Also removed the HEAD request sent before redirecting the user to the
GreaterWrong version of the page. While that was necessary in the past, GW
has largely reached feature parity with LessWrong, and so we don't need to
check if the page result in a HTTP 500. This fixes the issue where, sometimes,
the original page would load before the extension received the OK response from
GW and redirected.